requirements.txt
# ── Runtime ───────────────────────────────────────────────────────────────────
fastapi==0.134.0
uvicorn==0.41.0
starlette==0.52.1

# HTTP client (Mistral OCR + LLM calls)
httpx==0.28.1

# Settings / env
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.6.1
python-dotenv==1.1.0

# WebSocket proxy (ElevenLabs voice session)
websockets==15.0.1

# File uploads
python-multipart==0.0.22

# ── Indirect runtime dependencies ─────────────────────────────────────────────
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
h11==0.16.0
httpcore==1.0.9
idna==3.11
pillow==12.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0

# ── Dev / test only (not needed on the server) ────────────────────────────────
# pytest==9.0.2
# pytest-asyncio==1.3.0
# iniconfig==2.3.0
# packaging==26.0
# pluggy==1.6.0
# Pygments==2.19.2
